Expect NOT_SUPPORTED when attempting to derive a private key
Deriving a private key makes sense, but it is not currently supported
by the test suite. (As of version 1.0, the PSA Crypto specification
specifies how to do it for ECC keys but not for RSA keys.) Therefore
PSA_ERROR_NOT_SUPPORTED is more correct than
PSA_ERROR_INVALID_ARGUMENT, although it would also be fine for an
implementation to return PSA_SUCCESS (with the correct output), or
indeed INVALID_ARGUMENT if the implementation doesn't differentiate
betweent the key types it doesn't recognize.

(Deriving a _public_ key doesn't make sense, so there INVALID_ARGUMENT
is more correct.)