api-tests: Fix test case using length longer than valid buffer
The Crypto TEST 242 test-case for invalid signature length gives a
length field that is longer than the signature buffer length.
This causes a copy buffer overrun error that can be detected if the
non-secure application does not have read-access to this area.
This can trigger a PSA_ERROR_PROGRAMMER_ERROR return value instead
of the expected error code.

Make sure that the buffer is valid for the application so that the
test-case is actually verifying the functional test-case.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>